Sync full working state: pricing, search, and home categories
Backend:
- Fix admin dashboard 500: a best seller that has since been deactivated or
  deleted returned null and was dereferenced for its title. Titles now resolve
  in one batched query with a fallback, and summed money columns widen to long
  so a corrupt row cannot raise OverflowException.
- Market pricing moves to local currency. MerchantProduct gains PriceUsd as the
  dollar base; saving an exchange rate rewrites the derived MerchantPrice for
  dollar-quoted rows only, leaving locally priced merchants untouched. The rate
  itself moves to a language-neutral settings key so every culture resolves the
  same one.
- Search reaches non-restaurant merchants regardless of delivery coverage, and
  GetProductPrices ignores unpriced rows so an unstocked product can no longer
  win "best price" and blank a merchant out of the results.
- Home categories become a curated grid. Each tile carries an explicit
  destination (product category, merchant kind, merchant, or search term),
  which is what lets the apps route without matching words in a title.
- Restaurant categories link to real catalog categories and carry a live
  merchant count, so entries nobody sells in are withheld from customers.

Dashboard:
- New home categories page: add, reorder, activate, retarget and translate
  tiles, with inline validation of incomplete destinations.
- Restaurant categories gain the Home strip heading, a category picker showing
  merchant counts, and warnings for entries hidden from customers.

Mobile:
- Featured grid and the Home category strip render from backend configuration.
  Removes the title-matching router, the hardcoded cuisine list and its bundled
  images, and the merchant matcher whose fallback returned every merchant.

Also ignores dotnet publish output, which is regenerated per machine.

+++ jtak-backend-main/app/ApiModels/DashboardVm.cs
@@ -6,10 +6,10 @@
         public int UsersCount { get; set; }
         public int OrdersCount { get; set; }
         public int BillsCount { get; set; }
-        public int TotalOrdersValue { get; set; }
-        public int MerchantOrdersValue { get; set; }
-        public int JTakOrdersValue { get; set; }
-        public int JTakAdditionalOrdersValue { get; set; }
+        public long TotalOrdersValue { get; set; }
+        public long MerchantOrdersValue { get; set; }
+        public long JTakOrdersValue { get; set; }
+        public long JTakAdditionalOrdersValue { get; set; }
         public TopProduct[] TopProducts { get; set; }
     }
